// apps/backend/src/model/kanel/public/ProvisionedNewsFeedItem.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { NewsFeedItemId } from './NewsFeedItem';

/** Identifier type for public.ProvisionedNewsFeedItem */
export type ProvisionedNewsFeedItemPlatformId = string & { __brand: 'public.ProvisionedNewsFeedItem' };

/** Represents the table public.ProvisionedNewsFeedItem */
export default interface ProvisionedNewsFeedItem {
  news_feed_item_id: NewsFeedItemId;

  platform_id: ProvisionedNewsFeedItemPlatformId;
}

/** Represents the initializer for the table public.ProvisionedNewsFeedItem */
export interface ProvisionedNewsFeedItemInitializer {
  news_feed_item_id: NewsFeedItemId;

  platform_id: ProvisionedNewsFeedItemPlatformId;
}

/** Represents the mutator for the table public.ProvisionedNewsFeedItem */
export interface ProvisionedNewsFeedItemMutator {
  news_feed_item_id?: NewsFeedItemId;

  platform_id?: ProvisionedNewsFeedItemPlatformId;
}